fix: move token provider so it will be used only in lib's widget not in app
