+++ lib/lenra.dart
@@ -5,3 +5,4 @@ export 'package:lenra/view/lenra_route.dart';
 export 'package:lenra/model/token_provider.dart';
 export 'package:lenra/view/lenra_widget.dart';
 export 'package:lenra/oauth/oauth2_client.dart';
+export 'package:lenra/oauth/oauth2_widget.dart';
