+++ lib/oauth/oauth2_client.dart
@@ -1,12 +1,9 @@
 import 'dart:developer';
 
-import 'package:lenra/model/token_provider.dart';
 import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
 import 'package:oauth2_client/access_token_response.dart';
 import 'package:oauth2_client/oauth2_client.dart';
 import 'package:oauth2_client/oauth2_helper.dart';
-import 'package:provider/provider.dart';
 
 /// A class that represents a Lenra OAuth2 client.
 class LenraOauth2Client {
@@ -40,34 +37,5 @@ class LenraOauth2Client {
     // print(resp.body); but with logging driver
     log(resp!.accessToken!, name: 'lenra_oauth2_client');
     return resp;
-  }
-}
-
-/// A widget that handles the Lenra OAuth2 authentication flow.
-class LenraOauth2Widget extends StatelessWidget {
-  final Widget? child;
-
-  /// Creates a new instance of [LenraOauth2Widget].
-  const LenraOauth2Widget({Key? key, this.child}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    LenraOauth2Client oauth2 = LenraOauth2Client();
-
-    return FutureBuilder(
-      future: oauth2.refreshToken(),
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.done) {
-          if (snapshot.hasError) {
-            return Text('Error ${snapshot.error}');
-          }
-          context.read<TokenProvider>().token =
-              snapshot.data as AccessTokenResponse;
-          return child!;
-        } else {
-          return const CircularProgressIndicator();
-        }
-      },
-    );
   }
 }
